requirements.txt
asgiref==3.8.1
setuptools==75.6.0
Django>=5.1.3
djangorestframework==3.15.2
django-appconf==1.0.6
django-ckeditor-5==0.2.15
django-cleanup==9.0.0
django-imagekit==5.0.0
django-jazzmin==3.0.1
django-js-asset==2.2.0
pilkit==3.0
pillow==11.0.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
drf-yasg==1.21.8
django-filter==24.3
django-crispy-forms==2.3
djangorestframework-simplejwt==5.3.1
drf_rw_serializers>=1.4.0
drf-writable-nested>=0.7.1